requirements.txt
fastapi
uvicorn
deepface
opencv-python
numpy
pillow
python-multipart
tf-keras
tensorflow